==> src/02.feature_engineering.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from datetime import time

df = pd.read_excel('../dataset/BizPlay-2025_JAN.xlsx', sheet_name='2025년1월 거래데이터')
df

df.isnull().sum()

df['Sub-Merchant ID'] = df['Sub-Merchant ID'].fillna(0)
df

df.isnull().sum()

### Use Your Rules to Label Pseudo-Anomalies

outside_district = (
        (df['Customer Region (City/County/District)'].astype(str).str.strip() != df['Merchant Region (City/County/District)'].astype(str).str.strip()) &
        (df['Merchant Region (City/County/District)'] != 'ONLINE')
)
outside_district.sum()  # This tells you how many True rows

df['anomaly'] = 0
df.loc[outside_district, 'anomaly'] = 1

df['anomaly'].value_counts()

parsed_time = pd.to_datetime(df['Transaction Time'], format='%H%M%S').dt.time
rule_kids_after_10pm = (df['Customer Age'] < 8) & (parsed_time >= time(22, 0))
df.loc[rule_kids_after_10pm, 'anomaly'] = 1

df['anomaly'].value_counts()

late_night_mask = (parsed_time >= time(23, 0)) | (parsed_time < time(6, 0))
df.loc[late_night_mask, 'anomaly'] = 1

df['anomaly'].value_counts()

financial_aid_recipients_during_school_term = (parsed_time >= time(11, 0)) & (parsed_time < time(13, 0))
df.loc[financial_aid_recipients_during_school_term, 'anomaly'] = 1

df['anomaly'].value_counts()

from sklearn.preprocessing import LabelEncoder

for col in ['Customer Gender', 'Merchant ID', 'Sub-Merchant ID',
            'Transaction Processing Agency (VAN)']:
    df[col] = LabelEncoder().fit_transform(df[col].astype(str))

region_cols = ['Customer Region (City/County/District)', 'Merchant Region (City/County/District)']
region_df = df[region_cols].copy()

# Apply shared LabelEncoder
combined_region = pd.concat([
    region_df['Customer Region (City/County/District)'],
    region_df['Merchant Region (City/County/District)']
])

encoder = LabelEncoder()
encoder.fit(combined_region.astype(str))

df['Customer_Encoded'] = encoder.transform(region_df['Customer Region (City/County/District)'].astype(str))
df['Merchant_Encoded'] = encoder.transform(region_df['Merchant Region (City/County/District)'].astype(str))
df


df.to_csv("../dataset/cleaned_BizPlay_2025_JAN.csv", index=False)

df['anomaly'].value_counts()
